🔨 Refactor(#96): hook 분리
